requirements.txt
keyboard
sounddevice
numpy
openai-whisper
openai
edge-tts
pygame
pyyaml
python-dotenv
pyaudio
scipy